Molecular Hamiltonian (#321)
Make ElemCo an optional dependency through a package extension.
