+++ Project.toml
@@ -12,7 +12,6 @@ DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
 Distributed = "8ba89e20-285c-5b6f-9357-94700520ee1b"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
-ElemCo = "094d408e-8508-40f4-9646-a254980d91ac"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 Folds = "41a02a25-b8f0-4f67-bc48-60067656b558"
 HypergeometricFunctions = "34004b35-14d8-5ef3-9330-4cdb6864b03a"
@@ -48,11 +47,13 @@ VectorInterface = "409d34a3-91d5-4945-b6ec-7529ddf182d8"
 Arpack = "7d9fca2a-8960-54d3-9f78-7d1dccf2cb97"
 IterativeSolvers = "42fd0dbc-a981-5370-80f2-aaf504508153"
 KrylovKit = "0b1a1467-8014-51b9-945f-bf0ae24f4b77"
+ElemCo = "094d408e-8508-40f4-9646-a254980d91ac"
 
 [extensions]
 ArpackExt = "Arpack"
 IterativeSolversExt = "IterativeSolvers"
 KrylovKitExt = "KrylovKit"
+ElemCoExt = "ElemCo"
 
 [compat]
 Arpack = "0.5"
@@ -97,6 +98,7 @@ julia = "1.9"
 Arpack = "7d9fca2a-8960-54d3-9f78-7d1dccf2cb97"
 IterativeSolvers = "42fd0dbc-a981-5370-80f2-aaf504508153"
 KrylovKit = "0b1a1467-8014-51b9-945f-bf0ae24f4b77"
+ElemCo = "094d408e-8508-40f4-9646-a254980d91ac"
 
 [targets]
 test = ["Test"]
